+++ rust/sanctum-lst-list/tests/common/mod.rs
@@ -1,0 +1,1 @@
+pub const SOLANA_RPC_URL: &str = "https://api.mainnet-beta.solana.com";
