+++ rust/sanctum-lst-list/tests/mod.rs
@@ -1,0 +1,2 @@
+mod common;
+mod tests;
