+++ rust/sanctum-lst-list/tests/tests/mod.rs
@@ -1,0 +1,4 @@
+mod basic;
+mod logo;
+mod pool;
+mod sanctum_router;
